+++ src/scripts/modules/gooddata-writer-v3/routes.js
@@ -9,7 +9,7 @@ import loadType from './adapters/loadType';
 
 import DimensionsSection from './react/components/DimensionsSection';
 import dimensionsAdapter from './adapters/dimensions';
-import columnsEditorDefinition from './adapters/columnsEditorDefinition';
+import columnsEditorDefinition from './helpers/columnsEditorDefinition';
 import ToggleProjectAccess from './react/components/ToggleProjectAccess';
 import {CollapsibleSection} from '../configurations/utils/renderHelpers';
